requirements.txt
yacs
tyro
tqdm
nlopt
wandb
xacro
PyQt5
mplib
sapien
imgaug
pyyaml
PyOpenGL
tabulate
gymnasium
numpy==1.26.1
klampt==0.9.2
scikit-learn
opencv-python
pytransform3d
git+https://github.com/lixiny/manotorch.git
git+https://github.com/mattloper/chumpy.git
